Updated Image related functions
Edit.php - If the post has an image attached to it, a checkbox will appear allowing you to remove the image from the post.

Show.php - Image tag will only appear in the html if the post has an image attached to it.

Process.php - Added image resize functionality. Any images uploaded will be resized to 30% of their original size