Logout and CSS issues
Attempted to add a logout button. However, the user is logged out whenever the page is reloaded, even if they didn't click the logout button.

Also noticed an issue with the css page where the changes I make to the css page do not show up on the actual web page. But oddly changing the file name of the css file briefly fixes this.