Added hashed/salted passwords, and finished Admin CRUD
Admin can now create new users from within the users page.

When a new user is registered or created, or an existing user is edited through admin CRUD, their password is hashed and salted before being stored in the database. Now when a user attempts to login, their password is verified and compared to the hash stored in the database.