Completed user comments on posts
create.php - When creating a new post, I made the title and content fields required. And when a post is submitted to the process.php page, it will include the user_id of the person who made the post.

edit.php - Renamed the id variable to diorama_id to make it clearer what the variable referred to.

edit_user.php - Renamed the 'user' table to 'users'. And changed the input type of the password field from 'text' to 'password'.

index.php - Again, renamed id variable and gave it a clearer name.

process_users.php - Updated my sql queries to work with the renamed users table and user comments.

process.php - Removed the line that would check if title and content are at least 1 character long. Because it was causing problems with the user comments. And with the change to create.php (mentioned above), I felt it was no longer necessary. Also, like with edit.php I renamed variables to make it clearer what they refer to.

show.php - Renamed id variable to diorama_id. Added second query to handle user comments.

users.php - Updated the sql script to work with the changes mentioned above.