Added Category Sort
The category.php page I had uploaded before has been renamed to add_category.php. The category.php page will now be used to display a list of all posts in the chosen category.

index.php - Changed line 62 so that it uses the sanitized data instead of the unsanitized data.